requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==2.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.12.2
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
jsonfield==3.1.0
packaging==24.0
pluggy==1.5.0
pytest==8.1.1
pytest-django==4.8.0
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
